backend/requirements.txt
fastapi>=0.68.1,<0.69.0
uvicorn>=0.15.0,<0.16.0
httptools>=0.5.0
python-dotenv>=0.19.0,<0.20.0
openai>=0.27.0,<0.28.0
pydantic>=1.8.2,<2.0.0
requests>=2.26.0,<2.27.0
python-multipart>=0.0.5,<0.1.0
--find-links https://download.pytorch.org/whl/torch_stable.html
aiohttp>=3.8.1,<3.9.0
beautifulsoup4>=4.9.3,<4.10.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
watchfiles>=0.19.0
websockets>=10.4
h11>=0.14.0
